Refactor type-driven/contextual UI structure for DRY separation
