+++ lib/js/components/type-driven-ui-basics.mjs
@@ -17,6 +17,8 @@ import {
 import {
     _BaseContainerComponent
   , _BaseComponent
+  , _UIBaseList
+  , _UIBaseListContainerItem
   , UPDATE_STRATEGY
   , UPDATE_STRATEGY_NO_UPDATE
 } from './basics.mjs';
@@ -28,6 +30,10 @@ import {
 import {
     ProcessedPropertiesSystemMap
 } from './registered-properties-definitions.mjs';
+
+import {
+    getTransferTypesForModel
+} from './data-transfer-types.mjs';
 
 import {
      LineWidthLeadingModel
@@ -339,6 +345,90 @@ export function _BaseTypeDrivenContainerComponentMixin(Base) {
 
 // Backward compatible: the original class is just the mixin applied to _BaseContainerComponent
 export class _BaseTypeDrivenContainerComponent extends _BaseTypeDrivenContainerComponentMixin(_BaseContainerComponent) {
+}
+
+// UITypeDrivenListItem: a list item that renders type-driven widgets
+// for a fixed-type list model (e.g., CharGroupsListModel).
+export class UITypeDrivenListItem extends _BaseTypeDrivenContainerComponentMixin(_UIBaseListContainerItem) {
+    static TYPE_CLASS_PART = 'type_driven';
+
+    constructor(widgetBus, _zones, eventHandlers=[], draggable=false, injectable=null, transferTypePath=null) {
+        super(widgetBus, _zones, eventHandlers, draggable);
+        this._injectable = injectable;
+        this.ITEM_DATA_TRANSFER_TYPE_PATH = transferTypePath;
+
+        const entry = this.widgetBus.getEntry(this.widgetBus.rootPath)
+          , TypeClass = entry.constructor
+          , propertyRoot = this.widgetBus.rootPath.toString()
+          , ppsMap = ProcessedPropertiesSystemMap.fromPrefix(
+                propertyRoot, TypeClass.fields.keys())
+          , widgets = [
+                ...this._defineGenericWidgets(
+                      TypeClass
+                    , fieldName=>TypeClass.fields.has(fieldName)
+                    , {zone: 'local'}
+                    , ppsMap
+                    , this._injectable
+                )
+            ]
+          ;
+        this._initWidgets(widgets);
+    }
+}
+
+// UITypeDrivenList: a generic list container for fixed-type list models.
+// DnD types are derived from the item model via getTransferTypesForModel.
+export class UITypeDrivenList extends _UIBaseList {
+    static ROOT_CLASS = `ui_type_driven-list`;
+    static BASE_CLASSES = [...super.BASE_CLASSES, super.ROOT_CLASS];
+    static TYPE_CLASS_PART = null;
+    static UIItem = UITypeDrivenListItem;
+
+    constructor(widgetBus, _zones, injectable, label=null) {
+        const items = widgetBus.getEntry('collection')
+          , ItemModel = items.constructor.Model
+          , transferTypes = getTransferTypesForModel(ItemModel)
+          ;
+
+        const labelElement = label
+                ? widgetBus.domTool.createElement('span', {'class': 'typeroof-ui-label'}, label)
+                : null
+          , childrensMainZoneElement = widgetBus.domTool.createElement('div', {})
+          , localZoneElement = widgetBus.domTool.createElement('div', {}, [])
+          , zones = new Map([..._zones, ['local', localZoneElement], ['main', childrensMainZoneElement]])
+          ;
+        localZoneElement.append(childrensMainZoneElement);
+        super(widgetBus, zones, 'main');
+
+        if(label)
+            this.element.append(labelElement);
+        this.element.append(childrensMainZoneElement);
+
+        this._setClassesHelper([
+                ...(label ? [[labelElement, 'label']] : [])
+              , [childrensMainZoneElement, 'items']
+        ]);
+
+        this._injectable = injectable;
+        this._transferTypes = transferTypes;
+        this._ItemModel = ItemModel;
+        this.itemExtraArguments = [
+                this._injectable
+              , this._transferTypes.PATH
+        ];
+    }
+
+    get ITEM_DATA_TRANSFER_TYPE_PATH() {
+        return this._transferTypes.PATH;
+    }
+
+    get ITEM_DATA_TRANSFER_TYPE_CREATE() {
+        return this._transferTypes.CREATE;
+    }
+
+    _createNewItem(targetPath, insertPosition, items/*, value*/) {
+        return this._ItemModel.createPrimalDraft(items.dependencies);
+    }
 }
 
 export function createTypeToUIElementFunction(uiElementsMap
