+++ lib/js/components/type-driven-ui.mjs
@@ -13,7 +13,6 @@ import {
 import {
     _BaseContainerComponent
   , _BaseDynamicCollectionContainerComponent
-  , _UIBaseList
   , _UIBaseListContainerItem
 } from './basics.mjs';
 
@@ -35,6 +34,7 @@ import {
     require
   , _BaseTypeDrivenContainerComponentMixin
   , _BaseTypeDrivenContainerComponent
+  , UITypeDrivenList
   , createTypeToUIElementFunction
 } from './type-driven-ui-basics.mjs';
 
@@ -52,9 +52,6 @@ import {
   , createCharsSelector
 } from './actors/videoproof-contextual-models.mjs';
 
-import {
-    getTransferTypesForModel
-} from './data-transfer-types.mjs';
 
 import {
     OpenTypeFeaturesModel
@@ -66,6 +63,7 @@ import {
 
 import {
     UICharGroupContainer
+  , getCharGroupSummaryFromModel
 } from './ui-char-groups.mjs';
 
 import {
@@ -254,24 +252,6 @@ function _truncateSummary(value, maxLength = 36) {
     return `${text.slice(0, Math.max(0, maxLength - 1))}…`;
 }
 
-function _getCharGroupSummary(charGroupModel) {
-    const optionsModel = charGroupModel.get('options')
-      , options = optionsModel.isEmpty ? '(empty)' : optionsModel.value
-      , extendedModel = charGroupModel.get('extended')
-      , extended = !extendedModel.isEmpty && !!extendedModel.value
-      ;
-    if(options === 'custom') {
-        const customTextModel = charGroupModel.get('customText')
-          , customSeparatorModel = charGroupModel.get('customSeparator')
-          , customText = customTextModel.isEmpty ? '' : customTextModel.value
-          , customSeparator = customSeparatorModel.isEmpty ? '' : customSeparatorModel.value
-          , customSummary = _truncateSummary(customText, 24)
-          ;
-        return `custom: "${customSummary}"${customSeparator ? ` sep="${customSeparator}"` : ''}${extended ? ' +ext' : ''}`;
-    }
-    return `${options}${extended ? ' +ext' : ''}`;
-}
-
 function _getSelectorSummary(selectorModel) {
     const typeKey = selectorModel.get('selectorTypeKey').value;
     if(typeKey === ForeignKey.NULL)
@@ -289,7 +269,7 @@ function _getSelectorSummary(selectorModel) {
           , groups = []
           ;
         for(const [, charGroup] of charGroups)
-            groups.push(_getCharGroupSummary(charGroup));
+            groups.push(getCharGroupSummaryFromModel(charGroup));
         return `Simple(arg${argIndex}): ${groups.length ? groups.join(', ') : '(no groups)'}`;
     }
 
@@ -477,39 +457,6 @@ class UICharsSelectorContainer
     }
 }
 
-// UITypeDrivenListItem: a list item that renders type-driven widgets
-// for a fixed-type list model (e.g., CharGroupsListModel).
-// Combines _UIBaseListContainerItem (list lifecycle, drag-and-drop)
-// with _BaseTypeDrivenContainerComponentMixin (auto-widget generation).
-class UITypeDrivenListItem extends _BaseTypeDrivenContainerComponentMixin(_UIBaseListContainerItem) {
-    static TYPE_CLASS_PART = 'type_driven';
-
-    constructor(widgetBus, _zones, eventHandlers=[], draggable=false, injectable=null, transferTypePath=null) {
-        super(widgetBus, _zones, eventHandlers, draggable);
-        this._injectable = injectable;
-        this.ITEM_DATA_TRANSFER_TYPE_PATH = transferTypePath;
-
-        const entry = this.widgetBus.getEntry(this.widgetBus.rootPath)
-          , TypeClass = entry.constructor
-          , propertyRoot = this.widgetBus.rootPath.toString()
-          , ppsMap = ProcessedPropertiesSystemMap.fromPrefix(
-                propertyRoot, TypeClass.fields.keys())
-          , widgets = [
-                ...this._defineGenericWidgets(
-                      TypeClass
-                    , fieldName=>TypeClass.fields.has(fieldName)
-                    , {zone: 'local'}
-                    , ppsMap
-                    , this._injectable
-                )
-            ]
-          ;
-        this._initWidgets(widgets);
-    }
-}
-
-// UITypeDrivenList: a generic list container for fixed-type list models.
-// DnD types are derived from the item model via getTransferTypesForModel.
 class UIContextualTemplateRuleListItem extends _BaseTypeDrivenContainerComponentMixin(_UIBaseListContainerItem) {
     static TYPE_CLASS_PART = 'type_driven';
 
@@ -552,61 +499,6 @@ class UIContextualTemplateRuleListItem extends _BaseTypeDrivenContainerComponent
 
     _updateSummary() {
         this._summaryElement.textContent = _getTemplateRuleSummary(this.getEntry(this.widgetBus.rootPath));
-    }
-}
-
-class UITypeDrivenList extends _UIBaseList {
-    static ROOT_CLASS = `ui_type_driven-list`;
-    static BASE_CLASSES = [...super.BASE_CLASSES, super.ROOT_CLASS];
-    static TYPE_CLASS_PART = null;
-    static UIItem = UITypeDrivenListItem;
-
-    constructor(widgetBus, _zones, injectable, label=null) {
-        const items = widgetBus.getEntry('collection')
-          , ItemModel = items.constructor.Model
-          , transferTypes = getTransferTypesForModel(ItemModel)
-          ;
-
-        const labelElement = label
-                ? widgetBus.domTool.createElement('span', {'class': 'typeroof-ui-label'}, label)
-                : null
-          , childrensMainZoneElement = widgetBus.domTool.createElement('div', {})
-          , localZoneElement = widgetBus.domTool.createElement('div', {}, [])
-          , zones = new Map([..._zones, ['local', localZoneElement], ['main', childrensMainZoneElement]])
-          ;
-        localZoneElement.append(childrensMainZoneElement);
-        super(widgetBus, zones, 'main');
-
-        if(label)
-            this.element.append(labelElement);
-        this.element.append(childrensMainZoneElement);
-
-        this._setClassesHelper([
-                ...(label ? [[labelElement, 'label']] : [])
-              , [childrensMainZoneElement, 'items']
-        ]);
-
-        this._injectable = injectable;
-        this._transferTypes = transferTypes;
-        this._ItemModel = ItemModel;
-        // extra args: pass injectable and transferTypePath to the item
-        // used in super-class _createWrapper
-        this.itemExtraArguments = [
-                this._injectable
-              , this._transferTypes.PATH
-        ];
-    }
-
-    get ITEM_DATA_TRANSFER_TYPE_PATH() {
-        return this._transferTypes.PATH;
-    }
-
-    get ITEM_DATA_TRANSFER_TYPE_CREATE() {
-        return this._transferTypes.CREATE;
-    }
-
-    _createNewItem(targetPath, insertPosition, items/*, value*/) {
-        return this._ItemModel.createPrimalDraft(items.dependencies);
     }
 }
