+++ lib/js/components/ui-char-groups.mjs
@@ -527,6 +527,39 @@ export const
   ;
 
 
+function _truncateSummaryText(value, maxLength = 24) {
+    const text = `${value ?? ''}`;
+    if(text.length <= maxLength)
+        return text;
+    return `${text.slice(0, Math.max(0, maxLength - 1))}…`;
+}
+
+export function getCharGroupSummary(options, showExtended=false, customText='', customSeparator='') {
+    const option = options || '(empty)';
+    if(option === 'custom') {
+        const preview = _truncateSummaryText(customText, 24);
+        return `custom: "${preview}"${customSeparator ? ` sep="${customSeparator}"` : ''}${showExtended ? ' +ext' : ''}`;
+    }
+    return `${option}${showExtended ? ' +ext' : ''}`;
+}
+
+export function getCharGroupSummaryFromModel(charGroupModel) {
+    const optionsModel = charGroupModel.get('options')
+      , options = optionsModel.isEmpty ? '(empty)' : optionsModel.value
+      , extendedModel = charGroupModel.get('extended')
+      , showExtended = !extendedModel.isEmpty && !!extendedModel.value
+      ;
+    if(options === 'custom') {
+        const customTextModel = charGroupModel.get('customText')
+          , customSeparatorModel = charGroupModel.get('customSeparator')
+          , customText = customTextModel.isEmpty ? '' : customTextModel.value
+          , customSeparator = customSeparatorModel.isEmpty ? '' : customSeparatorModel.value
+          ;
+        return getCharGroupSummary(options, showExtended, customText, customSeparator);
+    }
+    return getCharGroupSummary(options, showExtended, '', '');
+}
+
 /* A helper element without DOM presence. */
 export class UIUpdateListener extends _BaseComponent {
     constructor(widgetBus, onUpdate) {
@@ -650,19 +683,11 @@ export class UICharGroupContainer extends _BaseTypeDrivenContainerComponent {
 
         const summaryWidget = this.getWidgetById('summary');
         if(summaryWidget && summaryWidget._ui && summaryWidget._ui._togglerElement) {
-            let summary = `${charGroup || '(empty)'}`;
-            if(charGroup === 'custom') {
-                const customText = this._getDefaults('customText', '')
-                  , customSeparator = this._getDefaults('customSeparator', '')
-                  , preview = customText.length > 24
-                        ? `${customText.slice(0, 23)}…`
-                        : customText
-                  ;
-                summary = `custom: "${preview}"${customSeparator ? ` sep="${customSeparator}"` : ''}`;
-            }
-            const showExtended = this._getDefaults('extended', false);
-            if(showExtended)
-                summary += ' +ext';
+            const customText = this._getDefaults('customText', '')
+              , customSeparator = this._getDefaults('customSeparator', '')
+              , showExtended = this._getDefaults('extended', false)
+              , summary = getCharGroupSummary(charGroup, showExtended, customText, customSeparator)
+              ;
             summaryWidget._ui._togglerElement.textContent = summary;
         }
     }
